// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/**
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/
export const UserRole = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserCityRole = {
  CITIZEN: 'CITIZEN',
  ADMIN: 'ADMIN'
} as const

export type UserCityRole = (typeof UserCityRole)[keyof typeof UserCityRole]


export const ComplaintCategory = {
  ILUMINACAO: 'ILUMINACAO',
  LIXO: 'LIXO',
  BURACO: 'BURACO',
  ESGOTO: 'ESGOTO',
  OUTROS: 'OUTROS'
} as const

export type ComplaintCategory = (typeof ComplaintCategory)[keyof typeof ComplaintCategory]


export const ComplaintStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
  REJECTED: 'REJECTED'
} as const

export type ComplaintStatus = (typeof ComplaintStatus)[keyof typeof ComplaintStatus]